Title: Implemented Feature to Add Songs to Custom Playlist
Description:

Created an addTrack method in the App component that accepts a track as an argument and adds it to the playlistTracks state if it's not already there.
Passed the addTrack method as a prop to the SearchResults component.
In the SearchResults component, passed the addTrack method as a prop to the TrackList component.
In the TrackList component, passed the addTrack method as a prop to each Track component.
In the Track component, added an event listener to the "add" button that triggers the addTrack method when clicked.
Ensured the Playlist component is rendered with the updated playlistTracks state whenever a track is added to the playlist.